total refactor of command handler, all commands, database queries

+++ commands/unpull.js
@@ -1,7 +1,8 @@
 const { SlashCommandBuilder, PermissionFlagsBits, EmbedBuilder, ButtonBuilder, ActionRowBuilder, ButtonStyle } = require('discord.js');
-const CONFIG = require('../models/config.js');
-const PULL = require('../models/pullrooms.js');
+const { getChannel, getMember } = require('../utils/fetch-utils.js');
 const superagent = require('superagent');
+const PULL = require('../models/pullrooms.js');
+const SV = require('../models/server-values.json');
 
 module.exports = {
     data: new SlashCommandBuilder()
@@ -15,24 +16,24 @@ module.exports = {
         ),
     async execute(interaction) {
         const idOption = interaction.options.getString('user-id'); // User ID to remove from pullroom
-        const cData = await CONFIG.findOne({ guildID: interaction.guild.id });                 // Get existing configuration data
+        const cData = interaction.client.cachedConfig; // Get existing configuration data
         const pData = await PULL.findOne({ guildID: interaction.guild.id, userID: idOption }); // Get existing pullroom data
 
         if (!cData) return interaction.reply({ content: 'I can\'t run that command if there is no data set up for the server! Use `/config-setup` first.' });
         if (!pData) return interaction.reply({ content: 'That user does not have a pullroom session open.' });
 
-        const USER_REPORTS_CID = '805795819722244148';                                             // Channel ID for the user reports channel
-        const ADMIN_DISCUSSION_CID = '781985975844733029';                                         // Channel ID for the admin channel
-        const log_sendChannelID = pData.admin ? ADMIN_DISCUSSION_CID : USER_REPORTS_CID;           // Log channel ID depending on admin value
-        const log_sendChannel = interaction.guild.channels.cache.get(log_sendChannelID);                  // Log channel object
-        const transcript_sendChannelID = pData.admin ? ADMIN_DISCUSSION_CID : cData.pulllogid; // Transcript channel ID depending on admin value
-        const transcript_sendChannel = interaction.guild.channels.cache.get(transcript_sendChannelID);    //  Transcript channel object
+        const USER_REPORTS_CID = SV.CHANNELS.USER_REPORTS;                                      // Channel ID for the user reports channel
+        const ADMIN_DISCUSSION_CID = SV.CHANNELS.ADMIN_DISCUSSION;                              // Channel ID for the admin channel
+        const log_sendChannelID = pData.admin ? ADMIN_DISCUSSION_CID : USER_REPORTS_CID;        // Log channel ID depending on admin value
+        const log_sendChannel = await getChannel(interaction.guild, log_sendChannelID);               // Log channel object
+        const transcript_sendChannelID = pData.admin ? ADMIN_DISCUSSION_CID : cData.pulllogid;    // Transcript channel ID depending on admin value
+        const transcript_sendChannel = await getChannel(interaction.guild, transcript_sendChannelID); //  Transcript channel object
 
-        const pullroomChannel = interaction.guild.channels.cache.get(pData.channelID); // Pullroom channel object
-        const pullMember = interaction.guild.members.cache.get(pData.userID);          // Member object for the pullroomed user
-
-        await interaction.deferReply();
+        const pullroomChannel = await getChannel(interaction.guild, pData.channelID); // Pullroom channel object
+        const pullMember = await getMember(interaction.guild, pData.userID);     // Member object for the pullroomed user
         if (pullMember) await pullMember.roles.remove(cData.pullroleid).catch(() => {}); // If the member is in the server, remove the role from them
+        
+        await interaction.deferReply();
 
         const transcriptEmbed = new EmbedBuilder()
             .setTitle(`${pData.userTag} (\`${pData.userID}\`)`)
@@ -82,10 +83,7 @@ module.exports = {
         if (pullroomChannel) await pullroomChannel.delete().catch(() => {});
         await interaction.followUp({ content: `Removed <@${pData.userID}> from their pullroom.` }).catch(() => {});
 
-        await PULL.findOneAndDelete({
-            guildID: interaction.guild.id,
-            userID: idOption
-        }).catch((err) => trailError(err));
+        await PULL.findOneAndDelete({ guildID: interaction.guild.id, userID: idOption }).catch((err) => trailError(err));
 
         const idUsername = interaction.client.users.cache.get(idOption)?.username;
         if (log_sendChannel) await log_sendChannel.send({ content: `🪢 <@${idOption}> ${idUsername ? `(${idUsername}) ` : ''}was removed from pullroom by ${interaction.user.username}` });
